+++ lib/pdfmd.rb
@@ -36,9 +36,7 @@
 # * Subject
 # * Keywords (optional)
 #
-# TODO: Include password protected PDF documents as well
 # TODO: Fix broken PDF files automatically
-# TODO: Enable logging in 'edit'
 # TODO: Read this: http://lostechies.com/derickbailey/2011/04/29/writing-a-thor-application/
 # TODO: ... and this: http://blog.paracode.com/2012/05/17/building-your-tools-with-thor/
 # gs \
@@ -59,7 +57,7 @@ require "i18n"
 require 'pathname'
 require 'logger'
 
-VERSION = '1.6.2'
+VERSION = '1.8.0'
 
 # Include general usage methods
 require_relative('pdfmd/methods.rb')
@@ -69,14 +67,13 @@ class DOC < Thor
   #
   # Show the current metadata tags
   #
-  # TODO: format output as JSON and YAML
   # TODO: Enable additional options
-  # TODO: Add command to show current settings (from hiera)
   #
   desc 'show', 'Show metadata of a file'
   method_option :all, :type => :boolean, :aliases => '-a', :desc => 'Show all metatags', :default => false, :required => false
   method_option :tag, :type => :string, :aliases => '-t', :desc => 'Show specific tag(s), comma separated', :required => false
-  #method_option :format, :type => :string, :aliases => '-f', :desc => 'Define output format', :required => false
+  method_option :format, :type => :string, :aliases => '-f', :desc => 'Define output format', :required => false
+  method_option :includepdf, :type => :boolean, :aliases => '-i', :desc => 'Include the filename in output', :required => false
   long_desc <<-LONGDESC
   == General
 
@@ -97,10 +94,35 @@ class DOC < Thor
 
   Relevant tags are Author,Creator, CreateDate, Title, Subject, Keywords.
 
+  This is the default action.
+
   --tag, -t
   \x5 Specify the metatag to show. The selected metatag must be one of the relevant tags. Other tags are ignored and nothing is returned.
 
   The value for the parameter is case insensitive: 'Author' == 'author'
+
+  Multiple Tags can be specificed, separated by a comma. 
+
+  If multiple tags are specified in a different order than the default order, the specified order will be used. This has an impact on the order of the fields when e.g. the output is exported in CSV format.
+
+  Hiera parameter: tag
+
+  --format, -f
+
+  Specify a different output format.
+
+  Available formats are json,yaml,csv,hash
+
+  The default output format is a human readable format.
+
+  Hiera parameter: format
+
+  --includepdf, -i
+
+  Include the filename of the PDF document in the output if this option is set to true.
+  Per default the filename is not included.
+
+  Hiera parameter: includepdf (boolean)
 
   == Example
 
@@ -116,12 +138,24 @@ class DOC < Thor
   # Show value for metatags 'Author','Title' for the file example.pdf
   \x5>CLI show -t author,title example.pdf
 
+  == Hiera
+
+  Here is an example configuration for hiera:
+
+  pdfmd::config
+    show:
+      format    : yaml
+      tag       : author,subject
+      includepdf: true
+
   LONGDESC
   def show(filename)
 
-    ENV['PDFMD_FILENAME'] = filename
-    ENV['PDFMD_TAGS']     = options[:tag]
-    ENV['PDFMD_ALL']      = options[:all].to_s
+    ENV['PDFMD_FILENAME']   = filename
+    ENV['PDFMD_TAGS']       = options[:tag]
+    ENV['PDFMD_ALL']        = options[:all].to_s
+    ENV['PDFMD_FORMAT']     = options[:format]
+    ENV['PDFMD_INCLUDEPDF'] = options[:includepdf].to_s
     require_relative('./pdfmd/show.rb')
 
   end
@@ -170,6 +204,8 @@ class DOC < Thor
   \x5 Rename file after updating the meta tag information according to the fields.
 
   This parameter is identical to running `> CLI rename <filename>`
+
+  Hiera parameter: rename
 
   General example:
 
@@ -236,9 +272,11 @@ class DOC < Thor
 
   See `> CLI help rename` for details about renaming.
 
+  To enable this feature in hiera add the key 'rename' into the section 'edit' with the value 'true'.
+
   LONGDESC
   method_option :tag, :type => :string, :aliases => '-t', :desc => 'Name of the Tag(s) to Edit', :default => false, :required => true
-  method_option :rename, :type => :boolean, :aliases => '-r', :desc => 'Rename file after changing meta-tags', :default => false, :required => false
+  method_option :rename, :type => :boolean, :aliases => '-r', :desc => 'Rename file after changing meta-tags', :required => false
   method_option :log, :aliases => '-l', :type => :boolean, :desc => 'Enable logging'
   method_option :logfile, :aliases => '-p', :type => :string, :desc => 'Define path to logfile'
   def edit(filename)