+++ lib/pdfmd/config.rb
@@ -11,9 +11,6 @@ require_relative '../string_extend.rb'
 require 'yaml'
 require 'pp'
 
-# TODO: the output can be probably made more pretty without
-# adding another requirement, can't it?
-
 #
 # If now options are set,
 # show the current settings
@@ -21,8 +18,8 @@ if opt_show.blank? or opt_show == 'true'
   opt_show = true
 end
 
+# Show the current settings
 case opt_show
-  # Show the current settings
 when true
 
   # As long as only Hiera is supported as external storage
@@ -32,9 +29,9 @@ when true
   puts ''
   hieraConfig = eval `hiera pdfmd::config`
   hieraConfig.sort.each do |key,value|
-    puts 'Command : ' + key
+    puts 'Command: ' + key
     puts value.to_yaml
-    puts "---\n\n"
+    puts ''
   end
 
 end